requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
python-dotenv==0.14.0
six==1.15.0
SQLAlchemy==1.3.19
Werkzeug==1.0.1
WTForms==2.3.3
psycopg2==2.8.5
flask-jwt-extended==3.24.1
flask-migrate==2.5.3
alembic==1.4.2
bcrypt==3.2.0
pylint==2.6.0
psycopg2-binary==2.8.5
boto3==1.14.56
flask-s3==0.3.3